Compare should respect the size of the data it compares.

+++ src/efibootmgr/efibootmgr.c
@@ -144,7 +144,7 @@ static int
 compare(const void *a, const void *b)
 {
 	int rc = -1;
-	uint32_t n1, n2;
+	uint16_t n1, n2;
 	memcpy(&n1, a, sizeof(n1));
 	memcpy(&n2, b, sizeof(n2));
 	if (n1 < n2) rc = -1;
